Big Updae
- Add card slug (add & edit)
- SEO Settings
- Google Analytics in User page
- Redirecting page as Tracker

+++ application/controllers/User.php
@@ -8,11 +8,14 @@ class User extends CI_Controller {
 		$this->load->library('form_validation');
 		if(!$this->session->userdata('ses_email')){
 			redirect('auth');
+			die();
 		}
 	} 
 
 	public function index(){
+		//set current page value
 		$data['page'] = 'link';
+		//getting user data by session
 		$data['user'] = $this->db->get_where('user', ['user_email' => $this->session->userdata('ses_email')])->row_array();
 
 		$this->load->view('templates/userpanel_header_v2', $data);
@@ -21,7 +24,9 @@ class User extends CI_Controller {
 	}
 
 	public function profile(){
+		//set current page value
 		$data['page'] = 'profile';
+		//getting user data by session
 		$data['user'] = $this->db->get_where('user', ['user_email' => $this->session->userdata('ses_email')])->row_array();
 		$data['social'] = $this->db->get_where('social', ['user_id' => $this->session->userdata('ses_id')])->row_array();
 		
@@ -64,17 +69,16 @@ class User extends CI_Controller {
 		}
 	}
 
-	public function setting($subpage){
+	public function setting($subpage=null){
 
 		if($subpage == 'theme'){
+			//set current page value
 			$data['page'] = 'setting';
+			//getting user data by session
 			$data['user'] = $this->db->get_where('user', ['user_email' => $this->session->userdata('ses_email')])->row_array();
-			
 			$data['cover'] = $this->db->get('cover')->result_array();
 			$data['layout'] = $this->db->get_where('layout', ['layout_status' => 1])->result_array();
 			$data['theme'] = $this->db->get('theme')->result_array();
-			//var_dump($data['cover']);
-			//die();
 
 			$this->form_validation->set_rules('setting-cover', 'Cover Image', 'trim|required');
 			$this->form_validation->set_rules('setting-layout', 'Layout', 'trim|required');
@@ -87,6 +91,7 @@ class User extends CI_Controller {
 				$chosen_cover = $this->input->post('setting-cover');
 				$chosen_layout = $this->input->post('setting-layout');
 
+				//Create whitelisting array items, so users can't input randomly
 				foreach($data['cover'] as $value_cover){
 					$avail_cover[] = $value_cover['cover_id'];
 				}
@@ -99,46 +104,88 @@ class User extends CI_Controller {
 				if (!in_array($chosen_layout, $avail_layout)){
 					$chosen_layout = 1;
 				}
+				//Create whitelisting array items, so users can't input randomly
 
-				$data_profile = [
+				$data_theme = [
 					'user_cover' => $chosen_cover,
 					'user_layout' => $chosen_layout,
 				];
 
-				if($this->db->update('user', $data_profile, array('user_id' => $this->session->userdata('ses_id')))){
+				if($this->db->update('user', $data_theme, array('user_id' => $this->session->userdata('ses_id')))){
 					$this->session->set_flashdata('message', '<div class="notification is-success">Settings Updated Successfully!</div>');
 					$error = $this->db->error();
-					redirect('user/setting');
+					redirect('user/setting/theme');
 				}else{
 					$error = $this->db->error();
 					$this->session->set_flashdata('message', '<div class="notification is-danger">Settings Update Failed!</div>');
-					redirect('user/setting');
+					redirect('user/setting/theme');
 				}
 			}
 		}else if($subpage == 'seo'){
+			//set current page value
 			$data['page'] = 'setting';
+			//getting user data by session
 			$data['user'] = $this->db->get_where('user', ['user_email' => $this->session->userdata('ses_email')])->row_array();
+			$data['seo'] = $this->db->get_where('seo', ['user_id' => $this->session->userdata('ses_id')])->row_array();
 
-			//var_dump($data['cover']);
-			//die();
-
-			$this->form_validation->set_rules('ga-tracking-id', 'GA Tracking ID', 'trim|required');
+			$this->form_validation->set_rules('meta-title', 'Meta Title', 'trim|min_length[10]|max_length[60]');
+			$this->form_validation->set_rules('meta-description', 'Meta Description', 'trim|max_length[160]');
+			$this->form_validation->set_rules('meta-keyword', 'Meta Keyword', 'trim|max_length[160]');
+			$this->form_validation->set_rules('meta-robot', 'Meta Robots', 'trim');
+			$this->form_validation->set_rules('meta-rating', 'Meta Rating', 'trim');
+			$this->form_validation->set_rules('ga-tracking-id', 'GA Tracking ID', 'trim');
 
 			if($this->form_validation->run() == false){
 				$this->load->view('templates/userpanel_header_v2', $data);
 				$this->load->view('user/settingb_v2', $data);
 				$this->load->view('templates/userpanel_footer_v2', $data);
 			}else{
+				$chosen_title = $this->input->post('meta-title');
+				$chosen_desc = $this->input->post('meta-description');
+				$chosen_keyword = $this->input->post('meta-keyword');
+				$chosen_robot = $this->input->post('meta-robot');
+				$chosen_rating = $this->input->post('meta-rating');
 				$chosen_gaid = $this->input->post('ga-tracking-id');
+
+				if($chosen_robots < 0 || $chosen_robots > 1){
+					$chosen_robots = 1;
+				}
+				if($chosen_rating < 0 || $chosen_rating > 1){
+					$chosen_rating = 0;
+				}
+
+				$data_meta = [
+					'meta_title' => $chosen_title,
+					'meta_description' => $chosen_desc,
+					'meta_keyword' => $chosen_keyword,
+					'meta_robot' => $chosen_robot,
+					'meta_rating' => $chosen_rating,
+					'gtag_id' => $chosen_gaid
+				];
+
+				if($this->db->update('seo', $data_meta, array('user_id' => $this->session->userdata('ses_id')))){
+					$this->session->set_flashdata('message', '<div class="notification is-success">Settings Updated Successfully!</div>');
+					$error = $this->db->error();
+					redirect('user/setting/seo');
+				}else{
+					$error = $this->db->error();
+					$this->session->set_flashdata('message', '<div class="notification is-danger">Settings Update Failed!</div>');
+					redirect('user/setting/seo');
+				}
 			}
+		}else{
+			redirect('user');
+			die();
 		}
 	}
 
 	public function addcard(){
+		//set current page value
 		$data['page'] = 'link';
+		//getting user data by session
 		$data['user'] = $this->db->get_where('user', ['user_email' => $this->session->userdata('ses_email')])->row_array();
 
-		$this->form_validation->set_rules('link-title', 'Title', 'trim|required|min_length[6]');
+		$this->form_validation->set_rules('link-title', 'Title', 'trim|required|min_length[6]|max_length[120]');
 		$this->form_validation->set_rules('link-destination', 'URL Destination', 'trim|required|valid_url');
 		$this->form_validation->set_rules('link-thumbnail', 'URL Thumbnail', 'trim|valid_url');
 
@@ -147,15 +194,24 @@ class User extends CI_Controller {
 			$this->load->view('user/add_card_v2', $data);
 			$this->load->view('templates/userpanel_footer_v2', $data);
 		}else{
-			//_addcards();
+			//Creating Card SLUG START
+			$next_id = $this->db->query("SHOW TABLE STATUS LIKE 'card'")->row(0)->Auto_increment;
+			$clean_title = preg_replace('/[^\p{L}\p{N}\s]/u', '', strtolower($this->input->post('link-title')));
+			$temp_title = str_replace(' ', '-', $clean_title);
+			$card_slug = $temp_title.'-'.$next_id;
+			//Creating Card SLUG END
+
+			//Collecting POST Value to Array
 			$data_card = [
 				'card_title' => $this->input->post('link-title'),
+				'card_slug' => $card_slug,
 				'card_url' => $this->input->post('link-destination'),
 				'card_thumbnail' => $this->input->post('link-thumbnail'),
 				'user_id' => $this->session->userdata('ses_id'),
 				'date_created' => time()
 			];
 
+			//Inserting Data to table
 			if($this->db->insert('card', $data_card)){
 				$this->session->set_flashdata('message', '<div class="notification is-success">New Link Added Successfully!</div>');
 				redirect('user');
@@ -167,12 +223,20 @@ class User extends CI_Controller {
 		}
 	}
 
-	public function editcard($card_id){
+	public function editcard($card_id=null){
+		//set current page value
 		$data['page'] = 'link';
+		//getting user data by session
 		$data['user'] = $this->db->get_where('user', ['user_email' => $this->session->userdata('ses_email')])->row_array();
 		$data['card'] = $this->db->get_where('card', ['card_id' => $card_id, 'user_id' => $this->session->userdata('ses_id')])->row_array();
 
-		$this->form_validation->set_rules('link-title', 'Title', 'trim|required|min_length[6]');
+		//if card content not found in db, then redirect
+		if(!$data['card']){
+			redirect('user');
+			die();
+		}
+
+		$this->form_validation->set_rules('link-title', 'Title', 'trim|required|min_length[6]|max_length[120]');
 		$this->form_validation->set_rules('link-destination', 'URL Destination', 'trim|required|valid_url');
 		$this->form_validation->set_rules('link-thumbnail', 'URL Thumbnail', 'trim|valid_url');
 
@@ -181,8 +245,22 @@ class User extends CI_Controller {
 			$this->load->view('user/edit_card_v2', $data);
 			$this->load->view('templates/userpanel_footer_v2', $data);
 		}else{
+			//Creating Card SLUG START
+			$clean_title = preg_replace('/[^\p{L}\p{N}\s]/u', '', strtolower($this->input->post('link-title')));
+			$arrayTitle = explode(" ",$clean_title);
+	
+			if(count($arrayTitle) < 6){
+				$temp_title = str_replace(' ', '-', $clean_title);
+			}else{
+				$temp_title = $arrayTitle[0].'-'.$arrayTitle[1].'-'.$arrayTitle[2].'-'.$arrayTitle[3].'-'.$arrayTitle[4];
+			}
+			
+			$card_slug = $temp_title.'-'.$card_id;
+			//Creating Card SLUG END
+
 			$data_card = [
 				'card_title' => $this->input->post('link-title'),
+				'card_slug' => $card_slug,
 				'card_url' => $this->input->post('link-destination'),
 				'card_thumbnail' => $this->input->post('link-thumbnail'),
 			];
@@ -195,18 +273,26 @@ class User extends CI_Controller {
 				$this->session->set_flashdata('message', '<div class="notification is-danger">Edit Link Failed!</div>');
 				redirect('user');
 			}
-			
-
 		}
 	}
 
-	public function deletecard($card_id){
+	public function deletecard($card_id=null){
+		//getting user data by session
+		$data['user'] = $this->db->get_where('user', ['user_email' => $this->session->userdata('ses_email')])->row_array();
+		$data['card'] = $this->db->get_where('card', ['card_id' => $card_id, 'user_id' => $this->session->userdata('ses_id')])->row_array();
+
+		//if card content not found in db, then redirect
+		if(!$data['card']){
+			redirect('user');
+			die();
+		}
+
 		if($this->db->delete('card', array('card_id' => $card_id, 'user_id' => $this->session->userdata('ses_id')))){
-			$this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">Card Deleted Successfully!</div>');
+			$this->session->set_flashdata('message', '<div class="notification is-success" role="alert">Card Deleted Successfully!</div>');
 			redirect('user');
 		}else{
 			$error = $this->db->error();
-			$this->session->set_flashdata('message', '<div class="alert alert-danger" role="alert">Deleting Card Failed!</div>');
+			$this->session->set_flashdata('message', '<div class="notification is-danger" role="alert">Deleting Card Failed!</div>');
 			redirect('user');
 		}
 	}
